third_party/WebKit/LayoutTests/http/tests/inspector/network/subresource-integrity-number-of-requests-for-script.html

<!DOCTYPE html>
<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
// Regression test for https://crbug.com/573269.
function loadIFrame() {
    var iframe = document.createElement('iframe');
    iframe.src = 'resources/call-success-with-integrity-frame.html';
    document.body.appendChild(iframe);
}

function test() {
    InspectorTest.addConsoleSniffer(step1);
    InspectorTest.evaluateInPage("loadIFrame()");

    function step1() {
        var requests = InspectorTest.findRequestsByURLPattern(/call-success.js/);
        InspectorTest.assertTrue(requests.length === 1);
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body>
<script>runTest();</script>
<p>Verify that only one request is made for basic script requests with integrity attribute.</p>
</body>
</html>
