third_party/WebKit/LayoutTests/http/tests/inspector/network/subresource-integrity-number-of-requests-for-script-expected.txt

CONSOLE MESSAGE: line 5: Script loaded successfully.
Verify that only one request is made for basic script requests with integrity attribute.
